api/ask/requirements.txt
langchain
openai
python-dotenv
sentence-transformers
faiss-cpu
langchain_community
azure-functions
numpy
sentence_transformers
fastapi
uvicorn
azure.core
azure-storage-blob


